pyproject.toml
[tool.poetry]
name = "supreme-umbrella"
version = "0.1.0"
description = "A homework solution for the OTUS 'Highload Architect' course."
authors = ["Mike <84048111+binhex386@users.noreply.github.com>"]
license = "Unlicense"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.0.2"
mysql-connector-python = "^8.0.28"
Flask-Login = "^0.5.0"
Flask-WTF = "^1.0.0"
email-validator = "^1.1.3"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
mypy = "^0.931"
flake8-bugbear = "^22.1.11"
flake8-bandit = "^2.1.2"

[tool.black]
target_version = ["py310"]

[tool.isort]
profile = "black"

[tool.mypy]
strict = true
disallow_untyped_decorators = false  # @with_appcontext
disallow_subclassing_any = false  # UserMixin

[[tool.mypy.overrides]]
module = [
    "flask_login",
    "flask_wtf",
    "mysql.*",
    "wtforms.*",
]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
